Work on loading/saving and syncing app preferences settings.

+++ src/Preferences.h
@@ -23,7 +23,7 @@ public:
 
 	BPath			fSettingsPath;
 
-	int8			fStartOfWeekOffset;
+	int32			fStartOfWeekOffset;
 	bool			fHeaderVisible;
 };
